Move replicator test
